Feat: 등록되어 있는 Event들의 수정 및 삭제 기능 추가 Feat: 공통적으로 사용할 메서드들을 추가할 CommonModule 추가 Refactor: class 명 rename 및 Adapter들의 코드 리팩터링

+++ app/src/main/java/com/zipdori/autoplanner/modules/calendarprovider/EventsVO.kt
@@ -1,15 +1,12 @@
 package com.zipdori.autoplanner.modules.calendarprovider
 
-import android.net.Uri
-import android.os.Parcel
 import android.os.Parcelable
-import androidx.versionedparcelable.VersionedParcelize
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
 class EventsVO(
     var id: Long,
-    var calendarId: Int,
+    var calendarId: Long,
     var organizer: String?,
     var title: String?,
     var eventLocation: String?,
